+++ db/migrate/20150612155140_add_duly_authorized_signature_print_to_agreements.rb
@@ -1,0 +1,9 @@
+class AddDulyAuthorizedSignaturePrintToAgreements < ActiveRecord::Migration
+  def change
+    add_column :agreements, :duly_authorized_representative_signature, :text
+    add_column :agreements, :duly_authorized_representative_signature_print, :string
+    add_column :agreements, :duly_authorized_representative_signature_date, :date
+    add_column :agreements, :duly_authorized_representative_token, :string
+    add_index  :agreements, :duly_authorized_representative_token, unique: true
+  end
+end