+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - **Agreement Changes**
   - The agreements administration is being merged into the reviewer page
     - This will help consolidate the DAUA approval process in a single place for principal reviewers
+  - Users not authorized to sign a DAUA are now provided a link that they can send to the authorized user
+  - An email is sent to the DAUA user when the authorized user has signed the DAUA
+  - Appending multiple users to an organization's DAUA has been removed
 - **Challenge Changes**
   - Added an export for the Flow Limitation challenge
 - **Tool Changes**